Stop using Babel
